+++ src/main/java/ru/aston/repository/impl/OrderRepositoryImpl.java
@@ -9,6 +9,7 @@ import ru.aston.repository.OrderRepository;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.Statement;
 
 public class OrderRepositoryImpl implements OrderRepository {
 
@@ -19,28 +20,34 @@ public class OrderRepositoryImpl implements OrderRepository {
     }
 
     @Override
-    public void createOrder(Order order) {
+    public Order createOrder(Order order) {
 
         String sqlQuery = "INSERT INTO order_t(name, user_id) VALUES(?, ?);";
 
+        long id = 0;
+
         try (Connection connection = dbManager.connect();
-             PreparedStatement stmt = connection.prepareStatement(sqlQuery)) {
+             PreparedStatement stmt = connection.prepareStatement(sqlQuery, Statement.RETURN_GENERATED_KEYS)) {
 
             stmt.setString(1, order.getName());
             stmt.setLong(2, order.getUserId());
 
             int affectedRows = stmt.executeUpdate();
 
-            if (affectedRows > 0) {
-                System.out.println("Order successfully create.");
-            } else {
-                System.out.println("Order doesn't create.");
+            if (affectedRows <= 0) {
+                throw new RuntimeException("Order doesn't create.");
+            }
+            ResultSet result = stmt.getGeneratedKeys();
+            if (result.next()) {
+                id = result.getLong(1);
+                order.setId(id);
+                return order;
             }
 
         } catch (Exception ex) {
-            System.out.println(ex.getMessage());
+            throw new RuntimeException(ex.getMessage());
         }
-
+        return order;
     }
 
     @Override
@@ -50,68 +57,66 @@ public class OrderRepositoryImpl implements OrderRepository {
                 "FROM order_t AS o " +
                 "WHERE id = ?;";
 
-        Order order = null;
+        Order order;
 
         try (Connection connect = dbManager.connect(); PreparedStatement stmt = connect.prepareStatement(sqlQuery)) {
             stmt.setLong(1, orderId);
             ResultSet result = stmt.executeQuery();
-
-            if (!result.first()) {
-                String.format("Order with id = %s was not found.", orderId);
-            }
-
             order = OrderMapper.orderMap(result);
-            return order;
 
         } catch (Exception ex) {
-            System.out.println(ex.getMessage());
+            throw new RuntimeException(ex.getMessage());
         }
-
-        return null;
+        return order;
     }
 
     @Override
-    public void updateOrder(Order order) {
+    public Order updateOrder(Order order) {
 
         String sqlQuery = "UPDATE order_t SET name = ? WHERE id = ?;";
 
+        long id = 0;
+
         try (Connection connect = dbManager.connect();
-             PreparedStatement stmt = connect.prepareStatement(sqlQuery)) {
+             PreparedStatement stmt = connect.prepareStatement(sqlQuery, Statement.RETURN_GENERATED_KEYS)) {
 
             stmt.setString(1, order.getName());
             stmt.setLong(2, order.getId());
+
             var affectedRows = stmt.executeUpdate();
 
-            if (affectedRows > 0) {
-                System.out.println("Order successfully update.");
-            } else {
-                System.out.println("Order doesn't update.");
+            if (affectedRows <= 0) {
+                throw new RuntimeException("Order doesn't update.");
             }
-
+            ResultSet result = stmt.getGeneratedKeys();
+            if (result.next()) {
+                id = result.getLong(1);
+                order.setId(id);
+                return order;
+            }
         } catch (Exception ex) {
-            System.out.println(ex.getMessage());
+            throw new RuntimeException(ex.getMessage());
         }
-
+        return order;
     }
 
     @Override
-    public void deleteOrderById(Long orderId) {
-        String sqlQuery = "DELETE FROM order_t WHERE id = ?;";
+    public void deleteOrderById(Long orderId, Long userId) {
+
+        String sqlQuery = "DELETE FROM order_t WHERE id = ? AND user_id = ?;";
 
         try (Connection connect = dbManager.connect();
              PreparedStatement stmt = connect.prepareStatement(sqlQuery)) {
 
             stmt.setLong(1, orderId);
+            stmt.setLong(2, userId);
             var affectedRows = stmt.executeUpdate();
 
-            if (affectedRows > 0) {
-                System.out.println("Order successfully delete.");
-            } else {
-                System.out.println("Order doesn't delete.");
+            if (affectedRows <= 0) {
+                throw new RuntimeException("Order doesn't delete.");
             }
-
         } catch (Exception ex) {
-            System.out.println(ex.getMessage());
+            throw new RuntimeException(ex.getMessage());
         }
     }
 }
